# docs/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tarcza Polska API",
        "description": "Backend for the Tarcza Polska civil resilience platform.\n\n* `/api/v1/*` - Citizen mobile app (Flutter). Authenticate with the JWT returned by `POST /api/v1/devices`.\n* `/api/command/*` - Command Center. Authenticate with the JWT returned by `POST /api/command/login`.\n\nAll spatial payloads are GeoJSON (WGS84, `[lng, lat]` order). Every error is an `ErrorResponse`.\n`GET /map`, `GET /verifications/pending` and `GET /alerts` send an `ETag`; repeat the value in\n`If-None-Match` to get `304 Not Modified`.\n",
        "version": "1.1.0"
    },
    "paths": {
...
        "/api/v1/verifications/pending": {
            "get": {
                "tags": [
                    "Verification"
                ],
                "summary": "Questions waiting for this device (poll on app foreground; pushes carry the same ids)",
                "operationId": "get_api_verification_pending",
                "responses": {
                    "200": {
                        "description": "List of pending questions. Sends an ETag.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/VerificationQuestion"
                                    }
                                }
                            }
                        }
                    },
                    "304": {
                        "description": "Not Modified (If-None-Match matched the ETag)"
                    },
                    "401": {
                        "description": "Missing, invalid or expired bearer token (error.code: unauthorized | token_expired)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/verifications/{id}": {
            "get": {
                "tags": [
                    "Verification"
                ],
                "summary": "One question (e.g. opened from a push)",
                "operationId": "get_api_verification_show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Question",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/VerificationQuestion"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Missing, invalid or expired bearer token (error.code: unauthorized | token_expired)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Unknown id, or a resource that belongs to another device (error.code: not_found)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/verifications/{id}/response": {
            "post": {
                "tags": [
                    "Verification"
                ],
                "summary": "Answer YES / NO / UNKNOWN",
                "operationId": "post_api_verification_respond",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RespondRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Recorded; returns the incident the answer fed into",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/VerificationResult"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Already answered (error.code: verification_already_answered)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "410": {
                        "description": "Question expired (error.code: verification_expired)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Missing, invalid or expired bearer token (error.code: unauthorized | token_expired)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Unknown id, or a resource that belongs to another device (error.code: not_found)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid body; error.violations[].field names the offending field (error.code: validation_failed)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "ErrorViolation": {
                "required": [
                    "field",
                    "message"
                ],
                "properties": {
                    "field": {
                        "description": "Name of the body field (maps 1:1 onto the form field)",
                        "type": "string"
                    },
                    "message": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "ErrorDetail": {
                "required": [
                    "code",
                    "message"
                ],
                "properties": {
                    "code": {
                        "description": "Stable machine-readable code: bad_request, unauthorized, token_expired, forbidden, not_found, conflict, gone, verification_already_answered, verification_expired, validation_failed, too_many_requests, internal_error.\n",
                        "type": "string"
                    },
                    "message": {
                        "type": "string"
                    },
                    "retryAfter": {
                        "description": "Seconds until the next attempt may succeed (429 only, also sent as the Retry-After header)",
                        "type": "integer"
                    },
                    "violations": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ErrorViolation"
                        }
                    }
                },
                "type": "object"
            },
            "ErrorResponse": {
                "required": [
                    "error"
                ],
                "properties": {
                    "error": {
                        "$ref": "#/components/schemas/ErrorDetail"
                    }
                },
                "type": "object"
            },
...
            "VerificationQuestion": {
                "required": [
                    "verificationId",
                    "incidentId",
                    "type",
                    "typeLabel",
                    "question",
                    "context",
                    "options",
                    "sentAt",
                    "expiresAt",
                    "answered"
                ],
                "properties": {
                    "verificationId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "incidentId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "type": {
                        "$ref": "#/components/schemas/ReportType"
                    },
                    "typeLabel": {
                        "type": "string"
                    },
                    "question": {
                        "type": "string"
                    },
                    "context": {
                        "type": "string"
                    },
                    "options": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/VerificationAnswer"
                        }
                    },
                    "sentAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "expiresAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "answered": {
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
            "VerificationResult": {
                "required": [
                    "verificationId",
                    "incidentId",
                    "thanks"
                ],
                "properties": {
                    "verificationId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "incidentId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "thanks": {
                        "description": "Polish thank-you line to show the user",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "ReportType": {
                "type": "string",
                "enum": [
                    "power_outage",
                    "water_outage",
                    "fuel_shortage",
                    "road_blocked",
                    "shelter_issue",
                    "other_threat"
                ]
            },
...
            "VerificationAnswer": {
                "type": "string",
                "enum": [
                    "yes",
                    "no",
                    "unknown"
                ]
            },
            "RespondRequest": {
                "required": [
                    "answer"
                ],
                "properties": {
                    "answer": {
                        "$ref": "#/components/schemas/VerificationAnswer"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "Bearer": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        {
            "Bearer": []
        }
    ],
...
}